[src/add_ui.c] Fixed unsigned → signed conversion of non-representable value (detected due to the new MPFR_SET_EXP assertion check). Now, 92 tests pass and 77 tests fail on Linux/x86_64.
[From SVN r8930 (branches/set_exp)]

+++ src/add_ui.c
@@ -35,7 +35,7 @@ mpfr_add_ui (mpfr_ptr y, mpfr_srcptr x, unsigned long int u, mpfr_rnd_t rnd_mode
     {
       mpfr_t uu;
       mp_limb_t up[1];
-      unsigned long cnt;
+      int cnt;
       int inex;
       MPFR_SAVE_EXPO_DECL (expo);
